+++ src/gen_enums.cpp
@@ -151,6 +151,7 @@ std::map<GenEnum::PropName, const char*> GenEnum::map_PropNames = {
     { prop_combined_xrc_file, "combined_xrc_file" },
     { prop_compiler_standard, "compiler_standard" },
     { prop_const_values, "generate_const_values" },
+    { prop_construction, "construction" },
     { prop_contents, "contents" },
     { prop_context_help, "context_help" },
     { prop_context_menu, "context_menu" },