+++ src/gen_enums.h
@@ -169,6 +169,7 @@ namespace GenEnum
         prop_combined_xrc_file,
         prop_compiler_standard,
         prop_const_values,
+        prop_construction,
         prop_contents,
         prop_context_help,
         prop_context_menu,