+++ src/panels/nav_panel.cpp
@@ -65,8 +65,7 @@ NavigationPanel::NavigationPanel(wxWindow* parent) : wxPanel(parent)
 {
     SetWindowStyle(wxBORDER_RAISED);
 
-    m_tree_ctrl = new wxTreeCtrl(this, wxID_ANY, wxDefaultPosition, wxDefaultSize,
-                                 wxTR_DEFAULT_STYLE | wxBORDER_SUNKEN);
+    m_tree_ctrl = new wxTreeCtrl(this, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxTR_DEFAULT_STYLE | wxBORDER_SUNKEN);
 
     int index = 0;
     wxSize gen_image_size = parent->FromDIP(wxSize(GenImageSize, GenImageSize));